Moved subscription tests to async daemon tests
